First implementation of MTD track extender for muons
