+++ RecoMTD/TrackExtender/plugins/BuildFile.xml
@@ -1,7 +1,8 @@
 <use name="RecoMTD/DetLayers"/>
 <use name="RecoMTD/Records"/>
+<use name="RecoMTD/TrackExtender"/>
 <use name="FWCore/Framework"/>
-<library file="TrackExtenderWithMTD.cc" name="RecoMTDTrackExtenderPlugins">
+<library file="*.cc" name="RecoMTDTrackExtenderPlugins">
   <use name="clhep"/>
   <use name="TrackingTools/TrackRefitter"/>
   <use name="TrackingTools/Records"/>
